+++ package.json
@@ -27,6 +27,7 @@
     "puppeteer-core": "^5.5.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "simple-icons": "^4.4.0",
     "twemoji": "^13.0.1"
   },
   "devDependencies": {
@@ -35,6 +36,7 @@
     "@types/puppeteer-core": "^5.4.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/simple-icons": "^2.5.0",
     "@types/twemoji": "^12.1.1",
     "prettier": "^2.2.1",
     "typescript": "^4.1.3"
